Corridor builder now behaves correctly.
Seems like most of the issues were caused with (int) NULL basically
evaluating to zero, which was unhelpful as zero was a valid value for
most of the checks. I've switch to using boolean values for ifs and -1
for initial values.

So, YAY the map generation code is correct. Now I just need to generate
some XML and send it to the client, and then I'll basically be back
where I was in python.